Made it so that, for a set of dummy tag links, selecting content on the grid-view page will display a union of the tagged products in the Select2 txt bar. It does not currently do the same for tagged pages, but that can easily be added once there is doe page data to work with.
Internally, I've basically added 3 components:
1) An array is kept and maintained of all currently selected item models
2) An abstract representation, in terms of Id's, is kept of what the txt boxes should be displaying. This might be unnecessary but I feel it makes things easier. This way you don't have to account for the delay there is when adding something to the txt box and have an easier time checking if something is present.
For this purpose I have also created a new local class called "FrequencyList" for lack of a better description which acts like a list where instead of duplicates, a quantity counter exists for every item.
3) A "TagData" object which contains a stripped down version of the tag linkages. Instead of linking object to object, it only links Id's. The data get's collected when the page first loads.
My reasoning behind this is that when an Ajax request is send either you have to halt progress until a reply is received, which isn't nice for the user, or you have to take into account that you can't be sure when certain data will arrive, which makes it difficult to program stuff. So, I thought it would be best if the most commonly used data got saved locally at first, and from there on ajax requests only need to be made for more major operations. Basically, saving the data locally prevents a ajax request from happening every time a content item is clicked, even when the tagged items are already all displayed.

All of this is within the "Index" class by the way.

Apart from that I also made a minor change to the template, giving a div element holding the two main select2 text boxes on the grid-view page the id "selection-edit"